Remove warnings: m8(m2)/m8baby boards. 1. Remove m8/m8m2/m8baby boards warnings.(not include customer folder) 2. Remove related driver warnings.(cmd_nvedit.c by Chunyu.Song)

+++ drivers/usb/gadget/v2_burning/aml_v2_burning.c
@@ -60,7 +60,7 @@ static unsigned _get_romcode_boot_id(void)
     unsigned* pBootId = (unsigned*)&(C_ROM_BOOT_DEBUG->boot_id);
     unsigned boot_id = *pBootId;
 #ifdef CONFIG_MESON_TRUSTZONE
-    boot_id = meson_trustzone_sram_read_reg32(pBootId);
+    boot_id = meson_trustzone_sram_read_reg32((uint32_t)pBootId);
 #endif// #ifdef CONFIG_MESON_TRUSTZONE
 
     return boot_id;
